requirements.txt
cassandra-driver==3.12.0
certifi==2017.11.5
chardet==3.0.4
futures==3.2.0
idna==2.6
kafka==1.3.5
ply==3.10
py-zipkin==0.9.0
py4j==0.10.4
pyspark==2.2.0
pytz==2017.3
requests==2.18.4
schedule==0.5.0
simplejson==3.13.2
six==1.11.0
style==1.1.0
thriftpy==0.3.9
update==0.0.1
urllib3==1.22
yahoo-finance==1.4.0
